% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dataset}
\alias{dataset}
\title{Initial Cleaning of our data}
\format{
## `dataset`
A data frame with 1,079 rows and 47 columns:
\describe{
  \item{TEAM}{Team name}
  \item{YEAR}{Year}
  ...
}
}
\usage{
dataset
}
\description{
A subset of data where we only kepy columns with numeric statistics that can
be used to weight each other and see which affects winning or losing the
most.
}
\keyword{datasets}
