% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_results}
\alias{example_results}
\title{Training dataset}
\format{
## `example_results`
A data frame with 63 rows and 5 columns:
\describe{
  \item{Team1}{Winnig team}
  \item{Team2}{Losing team}
  \item{Year}{Year}
  ...
}
}
\usage{
example_results
}
\description{
The results of the 2021 NCAA tournament that we can use to train our model
and see which statistics historically have impacted winning the most.
}
\keyword{datasets}
