% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_team_name.R
\name{format_team_name}
\alias{format_team_name}
\title{Format Team Name}
\usage{
format_team_name(team)
}
\arguments{
\item{team}{A character string representing the team name as inputted by the
user.}
}
\value{
Properly formatted teamname to fit the dataset.
}
\description{
Formats the team name as inputted by the user to account for slight errors
that may arrise from the user input.
}
\examples{
formatted_team <- format_team_name("utah st")
print(formatted_team)

}
