% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matchup_viz.R
\name{matchup_chart}
\alias{matchup_chart}
\title{Matchup Vis}
\usage{
matchup_chart(selected_team1, selected_team2)
}
\arguments{
\item{selected_team1}{a team_select object specifying the first team and of
interest.}

\item{selected_team2}{a team_select object specifying the second team and of
interest.}
}
\value{
The bar chart described above.
}
\description{
This function takes two teams and compares four important statistics
(Efficiency Field Goals, Free Throw Rate, Offensive Rebounds, and Turnover
Rate) to give insights on which team is better. The output of the function is
a bar chart showing side by side percentages for each team's metrics.
}
\examples{
matchup_chart(team_select("Baylor", 2023),
team_select("Purdue", 2023))

}
