% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matchup_prob.R
\name{matchup_prob}
\alias{matchup_prob}
\title{Matchup Probability Function
.
Gives two selectd teams a "power" rating based on their statistics and
uses this and a normal distribution to simulate
the probability of each team winning}
\usage{
matchup_prob(
  teamdata_matchup,
  n_sim = 10000,
  mean_diff = 57.049,
  sd_diff = 103.68
)
}
\arguments{
\item{teamdata_matchup}{A custom S3 object of class 'teamdata_matchup'
containing the subsetted data of two teams in the selected years.}
}
\value{
The "power" index of each team as well as the probability of each
team winning. Also a weighted coin flip if you want to select a winner from
the probabilities.
}
\description{
Matchup Probability Function
.
Gives two selectd teams a "power" rating based on their statistics and
uses this and a normal distribution to simulate
the probability of each team winning
}
\examples{
selected_matchup <- matchup_select("Purdue", 2024, "Utah St.", 2024)
matchup_prob(selected_matchup)

}
