% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matchup_select.R
\name{matchup_select}
\alias{matchup_select}
\title{Match-up Select}
\usage{
matchup_select(team1, year1, team2, year2)
}
\arguments{
\item{team1}{A character string representing the  first team name.}

\item{year1}{An integer representing the first year.}

\item{team2}{A character string representing the second team name.}

\item{year2}{An integer representing the second year.}
}
\value{
A custom S3 object of class 'teamdata_matchup' containing a
list of the subsetted data.
}
\description{
Creates a custom S3 object representing a list of data
for two specific teams and years.
}
