% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructors.R
\name{new_teamdata}
\alias{new_teamdata}
\title{Create New teamdata Object}
\usage{
new_teamdata(data_frame)
}
\arguments{
\item{data_frame}{A data frame containing the subsetted data.}
}
\value{
A custom S3 object of class 'teamdata' containing the subsetted data.
}
\description{
Creates an S3 object representing a subset of data for a
specific team and year.
}
