% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructors.R
\name{new_teamdata_matchup}
\alias{new_teamdata_matchup}
\title{Create New teamdata_matchup Object}
\usage{
new_teamdata_matchup(list_data)
}
\arguments{
\item{list_data}{A list containing the subsetted data.}
}
\value{
A custom S3 object of class 'teamdata_matchup'
containing the subsetted data.
}
\description{
Creates an S3 object representing a list of data for two
specific teams and years.
}
