% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/team_viz.R
\name{stat_table}
\alias{stat_table}
\title{Stat_table}
\usage{
stat_table(selected_team)
}
\arguments{
\item{selected_team}{a team_select object representing a dataframe for a team
in a given year.}
}
\value{
a data table with the statistics specified in the description.
}
\description{
This function creates a simple table showing important success indicators
for a team. These include efficiency field goals, free throw rate,
offensive rebounds, and turnover rate.
}
\examples{
stat_table(team_select("Baylor", 2023))

}
