% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/team_select.R
\name{team_select}
\alias{team_select}
\title{Team Select}
\usage{
team_select(team, year)
}
\arguments{
\item{team}{A character string representing the team name.}

\item{year}{An integer representing the year.}
}
\value{
A custom S3 object of class 'teamdata' containing the subsetted data.
}
\description{
Creates a custom S3 object representing a subset of data
for a specific team and year.
}
\examples{
selected_team <- team_select("Baylor", 2024)
print(selected_team)

}
