% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/team_viz.R
\name{team_table}
\alias{team_table}
\title{Team Vis Functions}
\usage{
team_table(selected_team)
}
\arguments{
\item{selected_team}{a team_select object representing a dataframe for a team
in a given year.}
}
\value{
a data table with team name, year, conference, seed, round, wins and
losses.
}
\description{
These functions create simple tables showing basic information about teams,
as well as important statistics that act as indicators of success.
}
\details{
Team_table

This function creates a simple table showing basic information about the team.
This includes team name, year, conference, seed, the round they made it to,
wins and losses.
}
\examples{
team_table(team_select("Baylor", 2023))

}
